+++ src/bookshelf/constants.py
@@ -3,7 +3,7 @@ Constants
 """
 import os
 
-DATA_FORMAT_VERSION = "v0.2.0"
+DATA_FORMAT_VERSION = "v0.2.1"
 ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
 
 DATA_DIR = os.path.join(ROOT_DIR, "data")