+++ tests/unit/test_notebook.py
@@ -93,7 +93,7 @@ def test_run_notebook(remote_bookshelf, name, version):
         )
 
         with pytest.raises(
-            ValueError, match=f"{os.path.join(td, version)} is not empty"
+            ValueError, match=re.escape(f"{os.path.join(td, version)} is not empty")
         ):
             run_notebook(
                 nb_name,