+++ CHANGELOG.rst
@@ -15,6 +15,11 @@ The changes listed in this file are categorised as follows:
 Unreleased
 ----------
 
+Added
+=====
+
+- (`!25 <https://gitlab.com/climate-resource/bookshelf/bookshelf/merge_requests/25>`_) Add primap-hist v2.4.1 and v2.4.2
+
 Changed
 =======