+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ The changes listed in this file are categorised as follows:
 Unreleased
 ----------
 
+- (`!28 <https://gitlab.com/climate-resource/bookshelf/bookshelf/-/merge_requests/28>`_) Fix file retrieval and publishing on windows
+
 Added
 =====