+++ CHANGELOG.rst
@@ -15,6 +15,10 @@ The changes listed in this file are categorised as follows:
 Unreleased
 ----------
 
+
+v0.2.2
+------
+
 Added
 =====